Moved commonly used function to utils.go.
